requirements.txt
asgiref==3.8.1
Django==5.1.5
django-debug-toolbar==5.0.1
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
style==1.1.0
tzdata==2025.1
update==0.0.1
whitenoise==6.8.2
